Split verifyOrAddFrontMatter in two
Split the yaml parsing and overwriting logic into a new function with a callback to do the modifying. This handles the yaml properly by parsing and then serialising it back.
The previous implementation broke if the key exists only as a nested key.
The value entry is now optional (defaulting to null)
The value entry is now any arbitrary value which is then serialised to yaml.
The only negative of this is that it populates null if no value is defined in the yaml