Added partial support to reposition pin
Added an attribute to the FileMarker class to store the inline match length
Added a context menu item to the map pin to edit location which stores the pin attribute
When the map is clicked and the pin attribute is not null it will rewrite the pin location with the clicked location.
So far implemented inline locations.
Still need to implement front matter locations.