Refactored getFileListByQuery and fixed #72
`this.app.metadataCache.getFileCache(file)[:].tags` only contained tags in the body of the file not files in the front matter.
The function getAllTags takes the file cache and returns a list of formatted tags from the front matter and the body.
Refactored the logic a bit to make it faster and easier to read.