Only parse out needed divicon properties once
Only parse out the background color and icon classes from the Leaflet div icon (used for resizable circle markers) once. This essentially caches these values and increases speed. However, this approach is still admittedly hacky and unwanted. As of now, I do not know of a way to get these values directly from Leaflet, which only exposes the content of the div as an HTMLElement. The original background color and icon class list are needed when creating a new div marker as a result of resizing. Future improvement on this method is desired.