Allow overwrite of an empty location property
The "Add geolocation (front matter) to current note" action correctly
avoids overwriting an existing location property. But I'm fussy about
the order of properties in my location notes, so I prefer to apply a
template with an empty location property in the right place. This
change allows the action to overwrite an existing entry if it is null.